Fix phase 0 training: AMP safety, torch.compile ordering, checkpoint state, quality sampling
Training pipeline (Adam):
- Reorder torch.compile to run after checkpoint load, preventing _orig_mod.*
  key prefix mismatches that crashed resume
- Save and restore scheduler (sched_g/sched_d) and AMP scaler (scaler_g/scaler_d)
  state in all checkpoints for full training-state fidelity on resume
- Integrate WeightedRandomSampler with quality_sampling config; falls back to
  plain shuffle when config key is absent (backward-compatible)
- Fix docstring: 192kHz → 96kHz throughout

AMP float16 safety (Adam):
- models/losses.py: cast y/y_hat to float32 before torch.stft and MelSpectrogram
  transforms (both reject fp16)
- models/mastering_losses.py: wrap EnCodec encoder calls in autocast(enabled=False)
  and cast inputs to float32 to prevent NaN under AMP (LSTM not fp16-safe)

GAN architecture (Kyle):
- models/generator.py: replace hardcoded 0.1 slope in HF branch with
  LEAKY_RELU_SLOPE constant; remove dead self.encoder ModuleList; move
  n_resblocks computation to __init__ as instance attribute to eliminate
  torch.compile graph break in forward()
- models/discriminator.py: replace all hardcoded 0.1 slopes with
  LEAKY_RELU_SLOPE constant; fix MSD forward() to use local y_d/y_hat_d
  loop variables instead of mutating the arguments in-place (torch.compile
  in-place mutation fix)

Dataset / config (Cain):
- data/dataset.py: add get_sample_weights() method returning per-file weights
  for WeightedRandomSampler; tiers with zero files receive weight 0
- configs/phase0.yaml: add quality_sampling block (weight_hires=7.0,
  weight_midres=2.0, weight_standard=1.0) for ~70/20/10 tier split

Docs:
- CLAUDE.md: add agent team roster and orchestration rules section

+++ models/losses.py
@@ -58,6 +58,10 @@ class STFTLoss(nn.Module):
             y = y.squeeze(1)
             y_hat = y_hat.squeeze(1)
 
+        # torch.stft does not support float16 — cast to float32 before calling.
+        y = y.float()
+        y_hat = y_hat.float()
+
         y_stft = torch.stft(y, self.fft_size, self.hop_size, self.win_size,
                             self.window, return_complex=True)
         y_hat_stft = torch.stft(y_hat, self.fft_size, self.hop_size, self.win_size,
@@ -124,8 +128,10 @@ class MelSpectrogramLoss(nn.Module):
             y = y.squeeze(1)
             y_hat = y_hat.squeeze(1)
 
-        y_mel = self.mel_transform(y)
-        y_hat_mel = self.mel_transform(y_hat)
+        # MelSpectrogram may produce dtype mismatches under AMP autocast —
+        # cast to float32 before calling the transform.
+        y_mel = self.mel_transform(y.float())
+        y_hat_mel = self.mel_transform(y_hat.float())
 
         return F.l1_loss(torch.log(y_hat_mel + 1e-8), torch.log(y_mel + 1e-8))
